+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > (P) - A premium version of that animation is available.
 
 ## [Unreleased]
-
+  - **New Assets**
+   - Added Last Day Dreaming Water, Fire, and Lightning Assets ([@ChasarooniZ](https://github.com/ChasarooniZ))
 ### Added
 
 - Items that grant effects (ex. spells) storing animations pertaining to those effects.